+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 * 1.0.0-next (@release_date@)
 
  * replaced `Assert` by webmozart/assert
+ * renamed `ResourceDiscovery::find()` to `findByType()`
+ * split `ResourceDiscovery::getBindings()` into `findByPath()` and
+   `getBindings()` without arguments
  
 * 1.0.0-beta2 (2015-01-27)